Remove lodash.get
This refactors the code base to get rid of lodash.get, which has been
deprecated in favor of the `?.` operator. Since this project promises
support with node >= 8.12 (see package.json:16) and `?.` is in Node.js
since v14 I do not use it in these changes. Rather, I use old school
JavaScript to achieve the same effect as these two alternatives. In most
cases where `lodash.get` was being used it was pretty much useless and
thus would not even require `?.`. In some cases it side-stepped the type
checker and the refactored code adds the appropriate type checks.